# Conflicts:
#	iOSClient/Main/Collection Common/NCCollectionViewDownloadThumbnail.swift
#	iOSClient/Main/Collection Common/Section Header Footer/NCSectionFirstHeader.swift
#	iOSClient/Media/NCMedia+CollectionViewDataSource.swift
#	iOSClient/Media/NCMedia+CollectionViewDelegate.swift
#	iOSClient/Media/NCMediaDownloadThumbnail.swift
#	iOSClient/Networking/E2EE/NCNetworkingE2EEDelete.swift
#	iOSClient/Networking/NCDownloadAction.swift
#	iOSClient/Networking/NCNetworking+Download.swift
#	iOSClient/Networking/NCNetworking+Synchronization.swift
#	iOSClient/Networking/NCNetworking+Task.swift
#	iOSClient/Networking/NCNetworking+Upload.swift
#	iOSClient/Networking/NCNetworking+WebDAV.swift
#	iOSClient/SceneDelegate.swift
#	iOSClient/Utility/NCUtilityFileSystem.swift
#	iOSClient/Viewer/NCViewerMedia/NCPlayer/NCPlayerToolBar.swift
#	iOSClient/Viewer/NCViewerMedia/NCViewerMedia.swift

+++ iOSClient/Networking/E2EE/NCNetworkingE2EEMarkFolder.swift
@@ -23,7 +23,9 @@ class NCNetworkingE2EEMarkFolder: NSObject {
 
         file.e2eEncrypted = true
 
-        let metadata = await self.database.addAndReturnMetadataAsync(self.database.convertFileToMetadata(file, isDirectoryE2EE: false))
+        guard let metadata = await self.database.addAndReturnMetadataAsync(self.database.convertFileToMetadata(file, isDirectoryE2EE: false)) else {
+            return .invalidData
+        }
 
         await self.database.addDirectoryAsync(e2eEncrypted: true, favorite: metadata.favorite, ocId: metadata.ocId, fileId: metadata.fileId, permissions: metadata.permissions, serverUrl: serverUrlFileName, account: metadata.account)
         await self.database.deleteE2eEncryptionAsync(predicate: NSPredicate(format: "account == %@ AND serverUrl == %@", metadata.account, serverUrlFileName))
