+++ iOSClient/Networking/NCNetworkingProcess.swift
@@ -29,28 +29,32 @@ actor NCNetworkingProcess {
         NotificationCenter.default.addObserver(forName: NSNotification.Name(rawValue: NCGlobal.shared.notificationCenterPlayerIsPlaying), object: nil, queue: nil) { [weak self] _ in
             guard let self else { return }
 
-            Task { await self.setScreenAwake(false) }
+            Task {
+                await self.setScreenAwake(false)
+            }
         }
 
         NotificationCenter.default.addObserver(forName: NSNotification.Name(rawValue: NCGlobal.shared.notificationCenterPlayerStoppedPlaying), object: nil, queue: nil) { [weak self] _ in
             guard let self else { return }
 
-            Task { await self.setScreenAwake(true) }
+            Task {
+                await self.setScreenAwake(true)
+            }
         }
 
         NotificationCenter.default.addObserver(forName: UIApplication.didEnterBackgroundNotification, object: nil, queue: nil) { [weak self] _ in
             guard let self else { return }
 
-            Task { await self.stopTimer() }
+            Task {
+                await self.stopTimer()
+            }
         }
 
         NotificationCenter.default.addObserver(forName: UIApplication.didBecomeActiveNotification, object: nil, queue: nil) { [weak self] _ in
             guard let self else { return }
 
             Task {
-                if !isAppInBackground {
-                    await self.startTimer(interval: self.maxInterval)
-                }
+                await self.startTimer(interval: self.maxInterval)
             }
         }
     }
@@ -64,6 +68,13 @@ actor NCNetworkingProcess {
     }
 
     func startTimer(interval: TimeInterval) async {
+        let isActive = await MainActor.run {
+            UIApplication.shared.applicationState == .active
+        }
+        guard isActive else {
+            return
+        }
+
         await stopTimer()
 
         lastUsedInterval = interval
@@ -72,7 +83,9 @@ actor NCNetworkingProcess {
 
         newTimer.setEventHandler { [weak self] in
             guard let self else { return }
-            Task { await self.handleTimerTick() }
+            Task {
+                await self.handleTimerTick()
+            }
         }
 
         timer = newTimer
@@ -103,13 +116,13 @@ actor NCNetworkingProcess {
             }
 
             let metadatas = await self.database.getMetadatasAsync(predicate: NSPredicate(format: "status != %d", self.global.metadataStatusNormal))
-            if let metadatas, !metadatas.isEmpty {
+            if !metadatas.isEmpty {
                 let tasks = await networking.getAllDataTask()
                 let hasSyncTask = tasks.contains { $0.taskDescription == global.taskDescriptionSynchronization }
-                let resultsTransfer = metadatas.filter { global.metadataStatusInTransfer.contains($0.status) }
+                let resultsScreenAwake = metadatas.filter { global.metadataStatusForScreenAwake.contains($0.status) }
 
                 if enableControllingScreenAwake {
-                    ScreenAwakeManager.shared.mode = resultsTransfer.isEmpty && !hasSyncTask ? .off : NCKeychain().screenAwakeMode
+                    ScreenAwakeManager.shared.mode = resultsScreenAwake.isEmpty && !hasSyncTask ? .off : NCKeychain().screenAwakeMode
                 }
 
                 await runMetadataPipelineAsync()
@@ -128,7 +141,6 @@ actor NCNetworkingProcess {
 
     private func removeUploadedAssetsIfNeeded() async {
         guard NCKeychain().removePhotoCameraRoll,
-              !isAppInBackground,
               let localIdentifiers = await self.database.getAssetLocalIdentifiersUploadedAsync(),
               !localIdentifiers.isEmpty else {
             return
@@ -149,15 +161,15 @@ actor NCNetworkingProcess {
 
     private func runMetadataPipelineAsync() async {
         let metadatas = await self.database.getMetadatasAsync(predicate: NSPredicate(format: "status != %d", self.global.metadataStatusNormal))
-        guard let metadatas, !metadatas.isEmpty else {
+        guard !metadatas.isEmpty else {
             return
         }
 
         /// ------------------------ WEBDAV
         let waitWebDav = metadatas.filter { self.global.metadataStatusWaitWebDav.contains($0.status) }
         if !waitWebDav.isEmpty {
-            let error = await metadataStatusWaitWebDav(metadatas: Array(waitWebDav))
-            if error != .success {
+            let (status, error) = await metadataStatusWaitWebDav(metadatas: Array(waitWebDav))
+            if  (error == .cancelled) || (status == global.metadataStatusWaitDelete && error != .success) {
                 return
             }
         }
@@ -190,7 +202,9 @@ actor NCNetworkingProcess {
         let isWiFi = self.networking.networkReachability == NKTypeReachability.reachableEthernetOrWiFi
         let sessionUploadSelectors = [self.global.selectorUploadFileNODelete, self.global.selectorUploadFile, self.global.selectorUploadAutoUpload]
         var counterUploading = metadatas.filter { $0.status == self.global.metadataStatusUploading }.count
-        for sessionSelector in sessionUploadSelectors where counterUploading < httpMaximumConnectionsPerHostInUpload {
+        for sessionSelector in sessionUploadSelectors {
+            guard counterUploading < httpMaximumConnectionsPerHostInUpload else { return }
+
             let limitUpload = max(0, httpMaximumConnectionsPerHostInUpload - counterUploading)
             let filteredUpload = metadatas
                 .filter { $0.sessionSelector == sessionSelector && $0.status == NCGlobal.shared.metadataStatusWaitUpload }
@@ -202,22 +216,26 @@ actor NCNetworkingProcess {
                 nkLog(debug: "PROCESS (UPLOAD) find \(metadatasWaitUpload.count) items")
             }
 
-            for metadata in metadatasWaitUpload where counterUploading < httpMaximumConnectionsPerHostInUpload {
+            for metadata in metadatasWaitUpload {
+                guard counterUploading < httpMaximumConnectionsPerHostInUpload else { return }
                 let metadatas = await NCCameraRoll().extractCameraRoll(from: metadata)
 
+                // no extract photo
                 if metadatas.isEmpty {
                     await self.database.deleteMetadataOcIdAsync(metadata.ocId)
                 }
 
-                for metadata in metadatas where counterUploading < httpMaximumConnectionsPerHostInUpload {
+                for metadata in metadatas {
+                    guard counterUploading < httpMaximumConnectionsPerHostInUpload,
+                          timer != nil else { return }
+
                     /// isE2EE
                     let isInDirectoryE2EE = metadata.isDirectoryE2EE
                     /// NO WiFi
                     if !isWiFi && metadata.session == networking.sessionUploadBackgroundWWan { continue }
-                    if isAppInBackground && (isInDirectoryE2EE || metadata.chunk > 0) { continue }
 
-                    await self.database.setMetadataStatusAsync(ocId: metadata.ocId,
-                                                               status: global.metadataStatusUploading)
+                    await self.database.setMetadataSessionAsync(ocId: metadata.ocId,
+                                                                status: global.metadataStatusUploading)
 
                     /// find controller
                     var controller: NCMainTabBarController?
@@ -271,12 +289,16 @@ actor NCNetworkingProcess {
         return
     }
 
-    private func metadataStatusWaitWebDav(metadatas: [tableMetadata]) async -> NKError {
+    private func metadataStatusWaitWebDav(metadatas: [tableMetadata]) async -> (status: Int?, error: NKError) {
 
         /// ------------------------ CREATE FOLDER
         ///
         let metadatasWaitCreateFolder = metadatas.filter { $0.status == global.metadataStatusWaitCreateFolder }.sorted { $0.serverUrl < $1.serverUrl }
         for metadata in metadatasWaitCreateFolder {
+            guard timer != nil else {
+                return (global.metadataStatusWaitCreateFolder, .cancelled)
+            }
+
             let resultsCreateFolder = await networking.createFolder(fileName: metadata.fileName,
                                                                   serverUrl: metadata.serverUrl,
                                                                   overwrite: true,
@@ -299,7 +321,7 @@ actor NCNetworkingProcess {
             }
 
             if resultsCreateFolder.error != .success {
-                return resultsCreateFolder.error
+                return (global.metadataStatusWaitCreateFolder, resultsCreateFolder.error)
             }
         }
 
@@ -307,8 +329,11 @@ actor NCNetworkingProcess {
         ///
         let metadatasWaitCopy = metadatas.filter { $0.status == global.metadataStatusWaitCopy }.sorted { $0.serverUrl < $1.serverUrl }
         for metadata in metadatasWaitCopy {
+            guard timer != nil else {
+                return (global.metadataStatusWaitCopy, .cancelled)
+            }
+
             let serverUrlTo = metadata.serverUrlTo
-            let serverUrlFileNameSource = metadata.serverUrl + "/" + metadata.fileName
             var serverUrlFileNameDestination = serverUrlTo + "/" + metadata.fileName
             let overwrite = (metadata.storeFlag as? NSString)?.boolValue ?? false
 
@@ -318,10 +343,10 @@ actor NCNetworkingProcess {
                 serverUrlFileNameDestination = serverUrlTo + "/" + fileNameCopy
             }
 
-            let resultCopy = await NextcloudKit.shared.copyFileOrFolderAsync(serverUrlFileNameSource: serverUrlFileNameSource, serverUrlFileNameDestination: serverUrlFileNameDestination, overwrite: overwrite, account: metadata.account)
+            let resultCopy = await NextcloudKit.shared.copyFileOrFolderAsync(serverUrlFileNameSource: metadata.serverUrlFileName, serverUrlFileNameDestination: serverUrlFileNameDestination, overwrite: overwrite, account: metadata.account)
 
-            await self.database.setMetadataStatusAsync(ocId: metadata.ocId,
-                                                       status: global.metadataStatusNormal)
+            await self.database.setMetadataSessionAsync(ocId: metadata.ocId,
+                                                        status: global.metadataStatusNormal)
 
             if resultCopy.error == .success {
                 let result = await NCNetworking.shared.readFileAsync(serverUrlFileName: serverUrlFileNameDestination, account: metadata.account)
@@ -335,7 +360,7 @@ actor NCNetworkingProcess {
             }
 
             if resultCopy.error != .success {
-                return resultCopy.error
+                return (global.metadataStatusWaitCopy, resultCopy.error)
             }
         }
 
@@ -343,15 +368,18 @@ actor NCNetworkingProcess {
         ///
         let metadatasWaitMove = metadatas.filter { $0.status == global.metadataStatusWaitMove }.sorted { $0.serverUrl < $1.serverUrl }
         for metadata in metadatasWaitMove {
+            guard timer != nil else {
+                return (global.metadataStatusWaitMove, .cancelled)
+            }
+
             let serverUrlTo = metadata.serverUrlTo
-            let serverUrlFileNameSource = metadata.serverUrl + "/" + metadata.fileName
             let serverUrlFileNameDestination = serverUrlTo + "/" + metadata.fileName
             let overwrite = (metadata.storeFlag as? NSString)?.boolValue ?? false
 
-            let resultMove = await NextcloudKit.shared.moveFileOrFolderAsync(serverUrlFileNameSource: serverUrlFileNameSource, serverUrlFileNameDestination: serverUrlFileNameDestination, overwrite: overwrite, account: metadata.account)
+            let resultMove = await NextcloudKit.shared.moveFileOrFolderAsync(serverUrlFileNameSource: metadata.serverUrlFileName, serverUrlFileNameDestination: serverUrlFileNameDestination, overwrite: overwrite, account: metadata.account)
 
-            await self.database.setMetadataStatusAsync(ocId: metadata.ocId,
-                                                       status: global.metadataStatusNormal)
+            await self.database.setMetadataSessionAsync(ocId: metadata.ocId,
+                                                        status: global.metadataStatusNormal)
 
             if resultMove.error == .success {
                 let result = await NCNetworking.shared.readFileAsync(serverUrlFileName: serverUrlFileNameDestination, account: metadata.account)
@@ -367,7 +395,7 @@ actor NCNetworkingProcess {
                     do {
                         try FileManager.default.removeItem(atPath: self.utilityFileSystem.getDirectoryProviderStorageOcId(metadata.ocId, userId: metadata.userId, urlBase: metadata.urlBase))
                     } catch { }
-                    await self.database.deleteVideoAsync(metadata: metadata)
+                    await self.database.deleteVideoAsync(metadata.ocId)
                     await self.database.deleteMetadataOcIdAsync(metadata.ocId)
                     await self.database.deleteLocalFileOcIdAsync(metadata.ocId)
                     // LIVE PHOTO
@@ -375,7 +403,7 @@ actor NCNetworkingProcess {
                         do {
                             try FileManager.default.removeItem(atPath: self.utilityFileSystem.getDirectoryProviderStorageOcId(metadataLive.ocId, userId: metadataLive.userId, urlBase: metadataLive.urlBase))
                         } catch { }
-                        await self.database.deleteVideoAsync(metadata: metadataLive)
+                        await self.database.deleteVideoAsync(metadataLive.ocId)
                         await self.database.deleteMetadataOcIdAsync(metadataLive.ocId)
                         await self.database.deleteLocalFileOcIdAsync(metadataLive.ocId)
                     }
@@ -387,7 +415,7 @@ actor NCNetworkingProcess {
             }
 
             if resultMove.error != .success {
-                return resultMove.error
+                return (global.metadataStatusWaitMove, resultMove.error)
             }
         }
 
@@ -395,6 +423,10 @@ actor NCNetworkingProcess {
         ///
         let metadatasWaitFavorite = metadatas.filter { $0.status == global.metadataStatusWaitFavorite }.sorted { $0.serverUrl < $1.serverUrl }
         for metadata in metadatasWaitFavorite {
+            guard timer != nil else {
+                return (global.metadataStatusWaitFavorite, .cancelled)
+            }
+
             let session = NCSession.Session(account: metadata.account, urlBase: metadata.urlBase, user: metadata.user, userId: metadata.userId)
             let fileName = utilityFileSystem.getFileNamePath(metadata.fileName, serverUrl: metadata.serverUrl, session: session)
             let resultsFavorite = await NextcloudKit.shared.setFavoriteAsync(fileName: fileName, favorite: metadata.favorite, account: metadata.account)
@@ -419,7 +451,7 @@ actor NCNetworkingProcess {
             }
 
             if resultsFavorite.error != .success {
-                return resultsFavorite.error
+                return (global.metadataStatusWaitFavorite, resultsFavorite.error)
             }
         }
 
@@ -427,12 +459,16 @@ actor NCNetworkingProcess {
         ///
         let metadatasWaitRename = metadatas.filter { $0.status == global.metadataStatusWaitRename }.sorted { $0.serverUrl < $1.serverUrl }
         for metadata in metadatasWaitRename {
-            let serverUrlFileNameSource = metadata.serveUrlFileName
+            guard timer != nil else {
+                return (global.metadataStatusWaitRename, .cancelled)
+            }
+
+            let serverUrlFileNameSource = metadata.serverUrlFileName
             let serverUrlFileNameDestination = metadata.serverUrl + "/" + metadata.fileName
             let resultRename = await NextcloudKit.shared.moveFileOrFolderAsync(serverUrlFileNameSource: serverUrlFileNameSource, serverUrlFileNameDestination: serverUrlFileNameDestination, overwrite: false, account: metadata.account)
 
             if resultRename.error == .success {
-                await self.database.setMetadataServeUrlFileNameStatusNormalAsync(ocId: metadata.ocId)
+                await self.database.setMetadataServerUrlFileNameStatusNormalAsync(ocId: metadata.ocId)
             } else {
                 await self.database.restoreMetadataFileNameAsync(ocId: metadata.ocId)
             }
@@ -444,7 +480,7 @@ actor NCNetworkingProcess {
             }
 
             if resultRename.error != .success {
-                return resultRename.error
+                return (global.metadataStatusWaitRename, resultRename.error)
             }
         }
 
@@ -456,11 +492,14 @@ actor NCNetworkingProcess {
             var returnError = NKError()
 
             for metadata in metadatasWaitDelete {
-                let serverUrlFileName = metadata.serverUrl + "/" + metadata.fileName
-                let resultDelete = await NextcloudKit.shared.deleteFileOrFolderAsync(serverUrlFileName: serverUrlFileName, account: metadata.account)
+                guard timer != nil else {
+                    return (global.metadataStatusWaitDelete, .cancelled)
+                }
 
-                await self.database.setMetadataStatusAsync(ocId: metadata.ocId,
-                                                           status: global.metadataStatusNormal)
+                let resultDelete = await NextcloudKit.shared.deleteFileOrFolderAsync(serverUrlFileName: metadata.serverUrlFileName, account: metadata.account)
+
+                await self.database.setMetadataSessionAsync(ocId: metadata.ocId,
+                                                            status: global.metadataStatusNormal)
 
                 if resultDelete.error == .success || resultDelete.error.errorCode == NCGlobal.shared.errorResourceNotFound {
                     do {
@@ -469,7 +508,7 @@ actor NCNetworkingProcess {
 
                     NCImageCache.shared.removeImageCache(ocIdPlusEtag: metadata.ocId + metadata.etag)
 
-                    await self.database.deleteVideoAsync(metadata: metadata)
+                    await self.database.deleteVideoAsync(metadata.ocId)
                     await self.database.deleteMetadataOcIdAsync(metadata.ocId)
                     await self.database.deleteLocalFileOcIdAsync(metadata.ocId)
 
@@ -479,9 +518,9 @@ actor NCNetworkingProcess {
                                                                                 account: metadata.account)
                     }
 
-                    metadatasError[metadata.detachedCopy()] = .success
+                    metadatasError[metadata] = .success
                 } else {
-                    metadatasError[metadata.detachedCopy()] = resultDelete.error
+                    metadatasError[metadata] = resultDelete.error
                     returnError = resultDelete.error
                 }
             }
@@ -492,10 +531,10 @@ actor NCNetworkingProcess {
             }
 
             if returnError != .success {
-                return returnError
+                return (global.metadataStatusWaitDelete, returnError)
             }
         }
 
-        return .success
+        return (nil, .success)
     }
 }
